Siehe Beschreibung für mehr.
- Aktualisiere die Speicher-Struktur.

- Handle Ergebnis und so im Core.

- Die Regeln können nun mit `X` nicht mehr angezeigt werden, falls dies zu nervig ist.

+++ common/include/lang.hpp
@@ -32,7 +32,7 @@
 
 namespace Lang {
 	std::string get(const std::string &key);
-	void load(int lang);
+	void load(uint8_t lang);
 };
 
 #endif
