+++ common/source/lang.cpp
@@ -28,7 +28,6 @@
 #include <stdio.h>
 
 static nlohmann::json appJson;
-
 #define LANG_PATH "romfs:/lang/"
 
 /*
@@ -39,7 +38,7 @@ static nlohmann::json appJson;
 std::string Lang::get(const std::string &key) {
 	if (!appJson.contains(key)) return "";
 
-	return appJson.at(key).get_ref<const std::string&>();
+	return appJson.at(key).get_ref<const std::string &>();
 }
 
 static std::string langs[] = { "de", "en" };
@@ -47,10 +46,10 @@ static std::string langs[] = { "de", "en" };
 /*
 	Lädt eine app.json datei.
 
-	int lang: Der Sprachen-Index, welcher oben als string array angegeben ist.
+	uint8_t lang: Der Sprachen-Index, welcher oben als string array angegeben ist.
 */
-void Lang::load(int lang) {
-	FILE* values;
+void Lang::load(uint8_t lang) {
+	FILE *values;
 	values = fopen((LANG_PATH + langs[lang] + "/app.json").c_str(), "rt");
 	if (values) appJson = nlohmann::json::parse(values, nullptr, false);
 	fclose(values);
